Clarify between slot and data usage of the same name ("items")
When first learning about slots and data tables, the examples given use the name
"items" for both the slot property names and the list of data to be displayed in
the data table. This can lead to confusion when your own data isn't called
"items". Renaming the data array from "items" to "desserts" will tease apart the
usage of those names and hopefully be clearer to the newer users.